# backend/internal/infrastructure/database/sqlc/storage/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.28.0

package storage

import (
	"github.com/google/uuid"
)

type Endpoint struct {
	ID           uuid.UUID
	ServiceID    uuid.UUID
	AssignedUser uuid.UUID
	Done         bool
	Name         string
	Path         string
	Method       string
	Description  string
	Details      []byte
}

type Project struct {
	ID        uuid.UUID
	Name      string
	CreatorID uuid.UUID
}

type ProjectMember struct {
	MemberID  uuid.UUID
	ProjectID uuid.UUID
}

type Service struct {
	ID             uuid.UUID
	ProjectID      uuid.UUID
	Name           string
	EndpointsTotal int32
	EndpointsDone  int32
}

type User struct {
	ID       uuid.UUID
	Name     string
	Password string
}